# Batch aggregator (experimental)