For changelog. Remove if this is non-significant change.

Category (leave one):
- Performance Improvement

Short description (up to few sentences):
Implemented batch variant of updating aggregate function states. It may lead to performance benefits.