@alexey-milovidov 2019-08-10
In extreme case it improves performance about 1.5 times:

```
SELECT SearchEngineID, count(), min(EventDate), max(EventDate) FROM hits_100m_single GROUP BY SearchEngineID ORDER BY count() DESC LIMIT 10
```

Before: 0.100 sec.
After: 0.061 sec.