@alexey-milovidov 2019-08-10
```
SELECT SearchEngineID, count(), min(EventDate), max(EventDate), avg(ResolutionWidth), sum(IsNotBounce), avgIf(SendTiming, SendTiming != -1) FROM hits_100m_single GROUP BY SearchEngineID ORDER BY count() DESC LIMIT 10
```

Before:

```
QPS: 4.730, RPS: 472976193.985, MiB/s: 4961.718, result RPS: 47.298, result MiB/s: 0.002.
0.000%  0.196 sec.
10.000% 0.198 sec.
20.000% 0.200 sec.
30.000% 0.202 sec.
40.000% 0.204 sec.
50.000% 0.206 sec.
60.000% 0.208 sec.
70.000% 0.211 sec.
80.000% 0.220 sec.
90.000% 0.227 sec.
95.000% 0.237 sec.
99.000% 0.268 sec.
99.900% 0.326 sec.
99.990% 0.332 sec.
```

After:

```
QPS: 6.023, RPS: 602252996.143, MiB/s: 6317.885, result RPS: 60.225, result MiB/s: 0.002.
0.000%  0.152 sec.
10.000% 0.155 sec.
20.000% 0.157 sec.
30.000% 0.158 sec.
40.000% 0.160 sec.
50.000% 0.162 sec.
60.000% 0.164 sec.
70.000% 0.165 sec.
80.000% 0.173 sec.
90.000% 0.181 sec.
95.000% 0.199 sec.
99.000% 0.204 sec.
99.900% 0.206 sec.
99.990% 0.206 sec.
```