@alexey-milovidov 2019-08-10
For the case without key, the performance advantage is even more obvious (4.5 times):

```
SELECT count(), min(EventDate), max(EventDate) FROM hits_100m_single

QPS: 13.025, RPS: 1302511482.858, MiB/s: 2484.343, result RPS: 13.025, result MiB/s: 0.000.
0.000%  0.067 sec.
10.000% 0.070 sec.
20.000% 0.071 sec.
30.000% 0.072 sec.
40.000% 0.074 sec.
50.000% 0.076 sec.
60.000% 0.077 sec.
70.000% 0.079 sec.
80.000% 0.081 sec.
90.000% 0.083 sec.
95.000% 0.088 sec.
99.000% 0.101 sec.
99.900% 0.106 sec.
99.990% 0.106 sec.

QPS: 55.848, RPS: 5584769904.231, MiB/s: 10652.103, result RPS: 55.848, result MiB/s: 0.001.
0.000%  0.015 sec.
10.000% 0.016 sec.
20.000% 0.016 sec.
30.000% 0.017 sec.
40.000% 0.017 sec.
50.000% 0.017 sec.
60.000% 0.018 sec.
70.000% 0.018 sec.
80.000% 0.018 sec.
90.000% 0.019 sec.
95.000% 0.020 sec.
99.000% 0.025 sec.
99.900% 0.087 sec.
99.990% 0.101 sec.
```