@alexey-milovidov 2019-08-11
Performance improvement is obvious from performance test.